requirements.txt
fastapi
uvicorn[standard]
pydantic
Pillow
PyPDF2
WeasyPrint
python-multipart